chore(refactor): Refactor Typeahead dropdown rendering for cleaner structure
Simplified the dropdown structure by moving inline styles and layout into a single wrapper `<div>`. This fixes the issue of the expanded maxHeight causing transparency issues. This occurs when a long list is rendered and the background only applies to the maxHeight, causing overlflowing elements to mix in with the UI in the layer below.